+++ icl/icl.hh
@@ -28,8 +28,8 @@ enum class Operation : uint8_t {
 };
 
 struct Request {
-  uint64_t id;
-  uint64_t sid;
+  uint32_t id;
+  uint32_t sid;
 
   Event eid;
   uint64_t data;
@@ -49,9 +49,9 @@ struct Request {
   uint8_t *buffer;
 
   Request();
-  Request(uint64_t, uint64_t, Event, uint64_t, Operation, LPN, uint32_t,
+  Request(uint32_t, uint32_t, Event, uint64_t, Operation, LPN, uint32_t,
           uint32_t, uint8_t *);
-  Request(uint64_t, Event, uint64_t, Operation, LPN, LPN);
+  Request(uint32_t, Event, uint64_t, Operation, LPN, LPN);
 };
 
 /**
